Remove manual check and refactor integration test
