+++ lib/spot-checker.js
@@ -33,14 +33,15 @@ export default class SpotChecker {
   }
 
   /**
-        *
-        * @param {object} args
-        * @param {RetrievalTask} args.task
-        * @param {object} args.stats
-        * @param {DagScope} args.dagScope
-        * @param {EntityBytesRange} args.entityBytesRange
-        * */
+          *
+          * @param {object} args
+          * @param {RetrievalTask} args.task
+          * @param {object} args.stats
+          * @param {DagScope} args.dagScope
+          * @param {EntityBytesRange} args.entityBytesRange
+          * */
   async executeSpotCheck ({ task, stats, dagScope, entityBytesRange }) {
+    console.log(`task ${task}`)
     console.log(`Calling Filecoin JSON-RPC to get PeerId of miner ${task.minerId}`)
     try {
       const peerId = await this.#getMinerPeerId(task.minerId)
@@ -84,14 +85,14 @@ export default class SpotChecker {
   }
 
   /**
-        * @param {object} args
-        * @param {RetrievalProtocol} args.protocol
-        * @param {string} args.address
-        * @param {string} args.cid
-        * @param {object} args.stats
-        * @param {DagScope} args.dagScope
-        * @param {EntityBytesRange} args.entityBytesRange
-        * */
+          * @param {object} args
+          * @param {RetrievalProtocol} args.protocol
+          * @param {string} args.address
+          * @param {string} args.cid
+          * @param {object} args.stats
+          * @param {DagScope} args.dagScope
+          * @param {EntityBytesRange} args.entityBytesRange
+          * */
   async fetchCAR ({ protocol, address, cid, stats, dagScope, entityBytesRange }) {
     // Abort if no progress was made for 60 seconds
     const controller = new AbortController()
@@ -112,8 +113,6 @@ export default class SpotChecker {
     const carBuffer = new ArrayBuffer(0, { maxByteLength: MAX_CAR_SIZE })
     const carBytes = new Uint8Array(carBuffer)
 
-    stats.startAt = new Date()
-
     try {
       const url = getRetrievalUrl({ protocol, address, cid, dagScope, entityBytesRange })
       console.log(`Fetching: ${url}`)
@@ -125,9 +124,6 @@ export default class SpotChecker {
       if (res.ok) {
         resetTimeout()
         for await (const value of res.body) {
-          if (stats.firstByteAt === null) {
-            stats.firstByteAt = new Date()
-          }
           stats.byteLength += value.byteLength
 
           // We want to limit how large content we are willing to download.
@@ -165,15 +161,13 @@ export default class SpotChecker {
     } finally {
       clearTimeout(timeout)
     }
-
-    stats.endAt = new Date()
   }
 
   /**
-        * @param {RetrievalTask} task
-        * @param {DagScope} dagScope
-        * @param {EntityBytesRange} [entityBytesRange]
-        * */
+          * @param {RetrievalTask} task
+          * @param {DagScope} dagScope
+          * @param {EntityBytesRange} [entityBytesRange]
+          * */
   async spotCheck (task, dagScope, entityBytesRange) {
     const stats = newStats()
     await this.executeSpotCheck({ task, stats, dagScope, entityBytesRange })
@@ -183,13 +177,13 @@ export default class SpotChecker {
   }
 
   /**
-        * @param {object} args
-        * @param {number} args.roundId
-        * @param {number} args.maxTasks
-        * @param {string} [args.minerId]
-        * @param {DagScope} [args.dagScope='entity']
-        * @param {EntityBytesRange} [args.entityBytesRange]
-    */
+          * @param {object} args
+          * @param {number} args.roundId
+          * @param {number} args.maxTasks
+          * @param {string} [args.minerId]
+          * @param {DagScope} [args.dagScope='entity']
+          * @param {EntityBytesRange} [args.entityBytesRange]
+      */
   async run ({ roundId, maxTasks, minerId, dagScope = 'entity', entityBytesRange }) {
     const results = []
     const retrievalTasks = await this.#tasker.getRetrievalTasks({ roundId, maxTasks, minerId, entityBytesRange })
@@ -214,9 +208,6 @@ export default class SpotChecker {
 export function newStats () {
   return {
     timeout: false,
-    startAt: null,
-    firstByteAt: null,
-    endAt: null,
     byteLength: 0,
     carChecksum: null,
     statusCode: null,
