+++ test.js
@@ -1,6 +1,5 @@
 import './test/ipni-client.test.js'
 import './test/miner-info.test.js'
 import './test/multiaddr.test.js'
-
 import './test/integration.js'
 import './test/spot-checker.js'
